Add general support for document cleanup
