+++ src/server/search-index/models.search-index.ts
@@ -5,11 +5,10 @@ import {
   MetricTimeframe,
   ModelHashType,
   ModelStatus,
-  PrismaClient,
   SearchIndexUpdateQueueAction,
 } from '@prisma/client';
 import { ModelFileType } from '~/server/common/constants';
-import { getOrCreateIndex } from '~/server/meilisearch/util';
+import { getOrCreateIndex, onSearchIndexDocumentsCleanup } from '~/server/meilisearch/util';
 import { EnqueuedTask } from 'meilisearch';
 import { getImagesForModelVersion } from '~/server/services/image.service';
 import { isDefined } from '~/utils/type-guards';
@@ -81,33 +80,6 @@ const onIndexSetup = async ({ indexName }: { indexName: string }) => {
   console.log('onIndexSetup :: updateRankingRulesTask created', updateRankingRulesTask);
 };
 
-const onDocumentsCleanup = async ({ db }: { db: PrismaClient }) => {
-  if (!client) {
-    return;
-  }
-
-  const queuedItemsToDelete = await db.searchIndexUpdateQueue.findMany({
-    select: {
-      id: true,
-    },
-    where: { type: INDEX_ID, action: SearchIndexUpdateQueueAction.Delete },
-  });
-
-  const itemIds = queuedItemsToDelete.map((queuedItem) => queuedItem.id);
-
-  // Only care for main index ID here. Technically, if this was working as a reset and using a SWAP,
-  // we wouldn't encounter delete items.
-  const index = await getOrCreateIndex(INDEX_ID);
-
-  if (!index) {
-    // If for some reason we don't get an index, abort the entire process
-    return;
-  }
-
-  const task = await index.deleteDocuments(itemIds);
-  await client.waitForTask(task.taskUid);
-};
-
 const onIndexUpdate = async ({
   db,
   lastUpdatedAt,
@@ -117,8 +89,11 @@ const onIndexUpdate = async ({
 
   // Confirm index setup & working:
   await onIndexSetup({ indexName });
+
   // Cleanup documents that require deletion:
-  await onDocumentsCleanup({ db });
+  // Always pass INDEX_ID here, not index name, as pending to delete will
+  // always use this name.
+  await onSearchIndexDocumentsCleanup({ db, indexName: INDEX_ID });
 
   let offset = 0;
   const modelTasks: EnqueuedTask[] = [];
