+++ src/server/search-index/tags.search-index.ts
@@ -1,5 +1,5 @@
 import { client } from '~/server/meilisearch/client';
-import { getOrCreateIndex } from '~/server/meilisearch/util';
+import { getOrCreateIndex, onSearchIndexDocumentsCleanup } from '~/server/meilisearch/util';
 import { EnqueuedTask } from 'meilisearch';
 import {
   createSearchIndexUpdateProcessor,
@@ -55,6 +55,10 @@ const onIndexUpdate = async ({ db, lastUpdatedAt, indexName }: SearchIndexRunCon
 
   // Confirm index setup & working:
   await onIndexSetup({ indexName });
+  // Cleanup documents that require deletion:
+  // Always pass INDEX_ID here, not index name, as pending to delete will
+  // always use this name.
+  await onSearchIndexDocumentsCleanup({ db, indexName: INDEX_ID });
 
   let offset = 0;
   const tagTasks: EnqueuedTask[] = [];
